+++ app/product/schemas.py
@@ -40,7 +40,7 @@ class ProductDiscountsCreate(BaseModel):
 class ProductBase(BaseModel):
     product_name: str = Field(min_length=1, description="商品名稱")
     description: str = Field(description="商品描述")
-    price: int = Field(gt=0, description="商品價格")
+    price: float = Field(gt=0, description="商品價格")
     one_set_price: Optional[int] = Field(ge=0, description="一組價格")
     one_set_quantity: Optional[int] = Field(ge=0, description="一組數量")
     stock_quantity: int = Field(ge=0, description="庫存數量")
